src/app/requirements.txt
Flask
pyfiglet
gitpython
flask-cors
gunicorn

